+++ frontend/src/redux/store.js
@@ -1,0 +1,8 @@
+import { configureStore } from "@reduxjs/toolkit";
+import userTypeReducer from "./reducers/userTypeSlice";
+
+export const store = configureStore({
+    reducer: {
+        userType: userTypeReducer,
+    }
+});
